swap test path
